trekk ut til queryBuilder.ts
